requirements.txt
beautifulsoup4==4.12.0
fastapi==0.95.0
httpx==0.23.3
pycryptodomex==3.17
pydantic==1.10.6
redis==4.5.2
requests==2.28.2
python-dotenv
